+++ docs/release-notes.md
@@ -11,6 +11,9 @@
 * For players:
   * Reduced memory usage.
 
+* For modders:
+  * Simplified tilesheet order warning added in 3.8.2.
+
 * For the Console Commands mod:
   * Removed experimental `performance` command. Unfortunately this impacted SMAPI's memory usage and the data was often misinterpreted. This may be replaced with more automatic performance alerts in a future version.